+++ packages/web-pkg/src/components/sideBar/Spaces/Details/SpaceDetails.vue
@@ -73,7 +73,7 @@ import { spaceRoleManager } from 'web-client/src/helpers/share'
 import { buildWebDavSpacesPath } from 'web-client/src/helpers'
 import { ImageDimension } from 'web-pkg/src/constants'
 import { useAccessToken, useStore } from 'web-pkg/src/composables'
-import SpaceQuota from '../../SpaceQuota.vue'
+import SpaceQuota from '../../../SpaceQuota.vue'
 import { formatDateFromISO } from 'web-pkg/src/helpers'
 import { configurationManager } from 'web-pkg/src/configuration'
 import { eventBus } from 'web-pkg/src/services/eventBus'
