+++ packages/web-pkg/src/components/sideBar/Spaces/Details/SpaceDetailsMultiple.vue
