+++ packages/web-pkg/src/components/sideBar/Spaces/SpaceNoSelection.vue
